seaparate split_data functionality and use it in preprocess functions
